pids: improve UI for optional DOI
