+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/Identifiers/PIDField/components/index.js
@@ -1,0 +1,14 @@
+// This file is part of Invenio-RDM-Records
+// Copyright (C) 2020-2023 CERN.
+// Copyright (C) 2020-2022 Northwestern University.
+//
+// Invenio-RDM-Records is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+export { ReservePIDBtn } from "./ReservePIDBtn";
+export { UnreservePIDBtn } from "./UnreservePIDBtn";
+export { ManagedUnmanagedSwitch } from "./ManagedUnmanagedSwitch";
+export { ManagedIdentifierCmp } from "./ManagedIdentifierCmp";
+export { UnmanagedIdentifierCmp } from "./UnmanagedIdentifierCmp";
+export { RequiredPIDField } from "../RequiredPIDField";
+export { OptionalDOIoptions } from "./OptionalDOIoptions";
