+++ invenio_rdm_records/services/components/pids.py
@@ -22,68 +22,38 @@ from ..errors import ValidationErrorWithMessageAsList
 from ..pids.tasks import register_or_update_pid
 
 
+def _get_optional_doi_transitions(record):
+    """Reusable method to validate optional DOI."""
+    RDM_OPTIONAL_DOI_TRANSITIONS = current_app.config["RDM_OPTIONAL_DOI_TRANSITIONS"]
+    if record:
+        record_pids = record.get("pids", {})
+        record_provider = record_pids.get("doi", {}).get("provider", "not_needed")
+        return RDM_OPTIONAL_DOI_TRANSITIONS.get(record_provider, {})
+    return {}
+
+
 class PIDsComponent(ServiceComponent):
     """Service component for PIDs."""
 
-    ALLOWED_DOI_PROVIDERS_TRANSITIONS = {
-        "datacite": {
-            "allowed_providers": ["datacite"],
-            "message": _(
-                "A previous version used a DOI registered from {sitename}. This version must also use a DOI from {sitename}."
-            ),
-        },
-        "external": {
-            "allowed_providers": ["external", "not_needed"],
-            "message": _(
-                "A previous version was published with a DOI from an external provider or without one. You cannot use a DOI registered from {sitename} for this version."
-            ),
-        },
-        "not_needed": {
-            "allowed_providers": ["external", "not_needed"],
-            "message": _(
-                "A previous version was published with a DOI from an external provider or without one. You cannot use a DOI registered from {sitename} for this version."
-            ),
-        },
-    }
-
-    def _validate_doi_transition(
-        self, new_provider, previous_published_provider, errors=None
-    ):
-        """If DOI is not required then we validate allowed DOI providers.
-
-        Each new version that is published must follow the ALLOWED_DOI_PROVIDERS_TRANSITIONS.
-        """
-        sitename = current_app.config.get("THEME_SITENAME", "this repository")
+    def _validate_optional_doi(self, record, previous_published_record, errors=None):
+        """.Validate optional DOI."""
         sitename = current_app.config.get("THEME_SITENAME", "this repository")
 
-        valid_transitions = self.ALLOWED_DOI_PROVIDERS_TRANSITIONS.get(
-            previous_published_provider, {}
-        )
-        if new_provider not in valid_transitions.get("allowed_providers", []):
-            error_message = {
-                "field": "pids.doi",
-                "messages": [
-                    valid_transitions.get("message").format(sitename=sitename)
-                ],
-            }
-
-            if errors is not None:
-                errors.append(error_message)
-            else:
-                raise ValidationErrorWithMessageAsList(message=[error_message])
-
-    def _validate_optional_doi(self, record, previous_published, errors=None):
-        """Reusable method to validate optional DOI."""
-        if previous_published:
-            previous_published_pids = previous_published.get("pids", {})
+        doi_transitions = _get_optional_doi_transitions(previous_published_record)
+        if doi_transitions:
             doi_pid = [pid for pid in record.pids.values() if "doi" in record.pids]
-            previous_published_provider = previous_published_pids.get("doi", {}).get(
-                "provider", "not_needed"
-            )
             new_provider = "not_needed" if not doi_pid else doi_pid[0]["provider"]
-            self._validate_doi_transition(
-                new_provider, previous_published_provider, errors
-            )
+            if new_provider not in doi_transitions.get("allowed_providers", []):
+                error_message = {
+                    "field": "pids.doi",
+                    "messages": [
+                        doi_transitions.get("message").format(sitename=sitename)
+                    ],
+                }
+                if errors is not None:
+                    errors.append(error_message)
+                else:
+                    raise ValidationErrorWithMessageAsList(message=[error_message])
 
     def create(self, identity, data=None, record=None, errors=None):
         """This method is called on draft creation.
@@ -110,10 +80,12 @@ class PIDsComponent(ServiceComponent):
         doi_required = "doi" in required_schemes
         can_manage_dois = self.service.check_permission(identity, "pid_manage")
         if not doi_required and not can_manage_dois:
-            previous_published = self.service.record_cls.get_latest_published_by_parent(
-                record.parent
+            previous_published_record = (
+                self.service.record_cls.get_latest_published_by_parent(record.parent)
             )
-            self._validate_optional_doi(record, previous_published, errors)
+            self._validate_optional_doi(
+                record, previous_published_record, errors=errors
+            )
 
         self.service.pids.pid_manager.validate(pids_data, record, errors)
         record.pids = pids_data
@@ -158,11 +130,10 @@ class PIDsComponent(ServiceComponent):
             # for any version of the record that will be published
             if draft.parent.get("pids", {}).get("doi"):
                 required_schemes.add("doi")
-
-            previous_published = (
+            previous_published_record = (
                 self.service.record_cls.get_previous_published_by_parent(record.parent)
             )
-            self._validate_optional_doi(draft, previous_published)
+            self._validate_optional_doi(draft, previous_published_record)
 
         self.service.pids.pid_manager.validate(draft_pids, draft, raise_errors=True)
 
@@ -257,12 +228,12 @@ class ParentPIDsComponent(ServiceComponent):
 
         # Check if a doi was added in the draft and create a parent DOI independently if
         # doi is required.
-        # Note: we don't have to check explicitely to the parent DOI creation only for
-        # datacite provider because we pass a `condition_func` below that it omits the
-        # minting if the pid selected is external
         if draft.get("pids", {}).get("doi"):
             required_schemes.add("doi")
 
+        # Note: we don't have explicitly to check for minting a parent DOI only for the
+        # managed provider because we pass a `condition_func` below that it omits the
+        # minting if the pid selected is external
         conditional_schemes = self.service.config.parent_pids_conditional
         for scheme in set(required_schemes):
             condition_func = conditional_schemes.get(scheme)
